[136] Generates new key pair

+++ lib/models/user/avatar/avatar_service.dart
@@ -139,13 +139,11 @@ abstract class AvatarService {
   }
 
   static Future<void> deleteAllAvatarsFromDeviceAndHive() async {
-    print('deleteAllAvatarsFromDeviceAndHive');
     final appDirectory = await getApplicationDocumentsDirectory();
-    print('appDirectory');
     final avatarsDirectory = Directory('${appDirectory.path}/avatars');
-    print('avatarsDirectory');
-    avatarsDirectory.deleteSync(recursive: true);
-    print('avatarsDirectory.deleteSync');
+    if (avatarsDirectory.existsSync()) {
+      avatarsDirectory.deleteSync(recursive: true);
+    }
     await AvatarHiveImage.cleanBox();
     log('deleted all avatar files and hive box clean');
   }
